Korrekturen

+++ mysql/MySQL-Dump.sql
@@ -451,7 +451,7 @@ DROP TABLE IF EXISTS `lernapp_SWPra`.`vorschlaege` ;
 
 CREATE TABLE IF NOT EXISTS `lernapp_SWPra`.`vorschlaege` (
   `id` INT NOT NULL,
-  `match_quote` INT NULL,
+  `match_quote` FLOAT NULL,
   `lernfaecher_id` VARCHAR(45) NULL,
   `main_person_id` VARCHAR(45) NULL,
   `match_profil_id` INT NOT NULL,
